# Improve aggregate_user performance on duplicates