### Description

<!-- What is the purpose of this PR? What is the current behavior? New behavior? Relevant links and/or information pertaining to PR? -->

Update the aggregate_user query to remove duplicates using UNION instead of UNION ALL. Not a huge improvement but this will reduce the number of changed_users which should make the upsert more efficient. 

### Tests

Manually tested against prod snapshot. 

Testing changed_users size.
Setting aggregate_user_latest_blocknumber to 1. 
w/o changes: changed_users produces 1429697 rows in 13 seconds.
w/ changes: changed_users produces 383275 rows in 8 seconds.

Testing upsert.
Setting aggregate_user_latest_blocknumber to 25000000. 
w/o changes: upsert affects 170523 rows in 57 seconds.
w/ changes: upsert affects 170523 rows in 49 seconds.


<!-- List the manual tests and repro instructions to verify that this PR works as anticipated. Include log analysis if possible. If this change impacts clients, make sure that you have tested the clients! -->

### How will this change be monitored?

index_aggregate_user duration logs
<!-- For features that are critical or could fail silently please describe the monitoring/alerting being added. -->


<!--
================ REMINDER: ================
If this PR touches a critical flow (such as Indexing, Uploads, Gateway or the Filesystem), make sure to add the `requires-special-attention` label.

** Add relevant labels as necessary. **
-->